+++ cmdbuild/src/main/java/org/cmdbuild/services/store/DashboardStore.java
@@ -6,8 +6,12 @@ import org.cmdbuild.model.dashboard.DashboardDefinition;
 
 public interface DashboardStore {
 	public Long create(DashboardDefinition dashboard);
+
 	public void update(Long dashboardId, DashboardDefinition dashboard);
+
 	public void delete(Long dashboardId);
+
 	public DashboardDefinition read(Long dashboardId);
-	public Map<Long, DashboardDefinition> list();
+
+	public Map<Integer, DashboardDefinition> list();
 }
